bugs fixed

+++ app/lib/drivers/interface/view/TypeUserOrderView.dart
@@ -234,7 +234,12 @@ class _LoginScreenState extends State<TypeOrderView> {
               color: theme.accentColor,
               onTap: () async{
                 //TODO:DELETE ORDER
-                Database().deleteOrder(order.data,(await UserBloc.of().outUser.first).model.id);
+                if(translateOrderCategory(order.data.state)!='Ordine Cancellato') Database().deleteOrder(order.data,(await UserBloc.of().outUser.first).model.id).then((value){
+                  Toast.show('Ordine cancellato!', context);
+                });
+                else {
+                  Toast.show('Ordine già cancellato!', context);
+                }
               },
             )
           ],
